+++ config/task.go
@@ -8,19 +8,19 @@ type TestCase struct {
 	Output string
 }
 
-// The type of the input, can be "stdin", "stdout" or maybe file-based
+// IOType represents the input format,it can be "stdin", "stdout" or maybe file-based
 // See Google Code Jam for file based IOTypes.
 type IOType struct {
 	Type string
 }
 
-// Specific to Java, Gives the name of the class and the computed task name.
+// LanguageDescription is specific to Java, Gives the name of the class and the computed task name.
 type LanguageDescription struct {
 	MainClass string
 	TaskClass string
 }
 
-// Task representation
+// Task representation.
 // Struct mapping to the json object received by competitive companion.
 type Task struct {
 	Name        string
@@ -28,7 +28,7 @@ type Task struct {
 	Url         string
 	Interactive bool
 	MemoryLimit float64
-	TimeLimit   float64
+	TimeLimit   int64
 	Tests       []TestCase
 	TestType    string
 	Input       IOType
